CMakeLists: cmake to build the gologpp package
package.xml: build, build export and exec dependencies for the gologpp package

example.gpp: contains the agent programm

src/agent_node: rosnode to start the gologpp-agent

The agent node loads the agent programm and parse it. The Pepper Backend gets initialized and then the programm gets invoked.

src/pepper_backend: The Pepper Backend is used to recieve the Transition from the gologpp agent programm and invokes the corresponding action on the robot

The Pepper Backend gets invoked in the main Loop of the Execution Context. The Execution Context passes the last Transition to the Pepper Backend. The Pepper Backend decides which action need to be executed based on the Actionname of current Transition. A template function gets called which sends the goal to the Actionserver. This template function is used cause the datatypes of the Actionclient and Goal contain the name of the corresponding ros package.
The template Function decides if the action is blocking or not. When it is blocking the node waits until the action is done or a timeout triggers. When it is not a blocking action the goal gets send to the server with a Callback-binding. The Callback-binding gets called when the action is done.
ToDo: The in-comming state of the actionserver need to be checked pass to the model where to decide which corresponding procedure need to be called. The parser need to bind the event procedure containing in the golog program with corresponding event function handler. The gologpp framework will handle the states succeed, preempted and failed.